Fix algorithms (and change the name of indexing to RoiSE)
